+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -8,13 +8,13 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class LottoTest {
+class LottoTicketTest {
 
-    Lotto lotto = new Lotto();
+    LottoTicket lottoTicket = new LottoTicket();
     List<Integer> list = new ArrayList<>();
     List<Integer> list2 = new ArrayList<>();
     List<Integer> target = new ArrayList<>();
-    LottoNumbers lottoNumbers;
+    Lotto lotto;
 
     @BeforeEach
     void setup() {
@@ -24,19 +24,19 @@ class LottoTest {
         list.add(4);
         list.add(5);
         list.add(6);
-        lotto.addLottoNumbers(new LottoNumbers(list));
+        lottoTicket.addLottoNumbers(new Lotto(list));
 
         list2.add(2);
         list2.add(3);
         list2.add(5);
-        lotto.addLottoNumbers(new LottoNumbers(list2));
+        lottoTicket.addLottoNumbers(new Lotto(list2));
 
         target.add(1);
         target.add(2);
         target.add(3);
         target.add(4);
         target.add(5);
-        lottoNumbers = new LottoNumbers(target);
+        lotto = new Lotto(target);
     }
 
     @Test
@@ -45,11 +45,11 @@ class LottoTest {
         matchOf.add(Match.getMatchInfo(5));
         matchOf.add(Match.getMatchInfo(3));
 
-        assertThat(lotto.getMatch(lottoNumbers.convertLottoNumbers())).isEqualTo(matchOf);
+        assertThat(lottoTicket.getMatch(lotto.convertLottoNumbers())).isEqualTo(matchOf);
     }
 
     @Test
     void matchLottoCount_test() {
-        assertThat(lotto.getMatchLottoCount(target).size()).isEqualTo(2);
+        assertThat(lottoTicket.getMatchLottoCount(target).size()).isEqualTo(2);
     }
 }
